% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{Parasite abundance}
\usage{
abundance(x)
}
\arguments{
\item{x}{number of parasites by host}
}
\value{
data frame with Mean Abundance and Standard Deviation.
}
\description{
Abundance of parasites according to Bush et al. (1997).
}
\examples{
attach(birdticks)
abundance(ticks)
}
\references{
Bush, Albert O., Kevin D. Lafferty, Jeffrey M. Lotz, and Allen W. Shostak. 1997. “Parasitology Meets Ecology on Its Own Terms: Margolis et al. Revisited.” The Journal of Parasitology 83 (4): 575. \doi{10.2307/3284227}
}
