% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdticks}
\alias{birdticks}
\title{Number of ticks by bird}
\format{
A data frame with 589 rows and 2 variables:
\describe{
\item{birds}{bird species}
\item{ticks}{number of ticks}
...
}
}
\usage{
birdticks
}
\description{
Number of ticks by bird
}
\keyword{datasets}
