% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buck_vapour_water}
\alias{buck_vapour_water}
\title{water vapour pressure
Buck equation for vapour pressure of water in mmhg for \code{\link{saturation_deficit_bulb}} and \code{\link{relative_humidity_bulb}}}
\usage{
buck_vapour_water(t)
}
\arguments{
\item{t}{temperature in Celsius degree}
}
\description{
water vapour pressure
Buck equation for vapour pressure of water in mmhg for \code{\link{saturation_deficit_bulb}} and \code{\link{relative_humidity_bulb}}
}
\keyword{internal}
