% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrepancy.R
\name{discrepancy}
\alias{discrepancy}
\title{Discrepancy index}
\usage{
discrepancy(x)
}
\arguments{
\item{x}{number of parasites by host}
}
\description{
Aggregation Discrepance index according to Poulin (1993).
}
\examples{
attach(birdticks)
discrepancy(ticks)
}
\references{
Poulin, Robert. 1993. “The Disparity between Observed and Uniform Distributions: A New Look at Parasite Aggregation.” International Journal for Parasitology 23 (7): 937–44. \doi{10.1016/0020-7519(93)90060-C}
}
