% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity}
\alias{intensity}
\title{Parasite intensity}
\usage{
intensity(x)
}
\arguments{
\item{x}{number of parasites by host}
}
\value{
data frame with Mean Intensity and Standard Deviation.
}
\description{
Intensity of parasites according to Bush et al. (1997).
}
\examples{
attach(birdticks)
intensity(ticks)
}
\references{
Bush, Albert O., Kevin D. Lafferty, Jeffrey M. Lotz, and Allen W. Shostak. 1997. “Parasitology Meets Ecology on Its Own Terms: Margolis et al. Revisited.” The Journal of Parasitology 83 (4): 575. \doi{10.2307/3284227}
}
