% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Test if x is a integer number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{number to be tested}
}
\description{
Test if x is a integer number
}
\note{
function derived from example of \code{\link[base]{integer}} documentation.
}
\keyword{internal}
