% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{k_negative_binomial}
\alias{k_negative_binomial}
\alias{k_negative_binomial_inverse}
\title{Negative binomial}
\usage{
k_negative_binomial(x)

k_negative_binomial_inverse(x)
}
\arguments{
\item{x}{number of parasites by host}
}
\description{
Calculate binomial negative aggregation index. Note that this function is a very simple implementation of \code{\link[fitdistrplus]{fitdist}}  from fitdistrplus package for pragmatic purpose. You can do much more with fitdistrplus.
}
\section{Functions}{
\itemize{
\item \code{k_negative_binomial_inverse()}: K from binomial negative inverse

}}
\examples{
attach(birdticks)
k_negative_binomial(ticks)
attach(birdticks)
k_negative_binomial_inverse(ticks)
}
\references{
Delignette-Muller, Marie Laure, and Christophe Dutang. 2015. “Fitdistrplus: An R Package for Fitting Distributions.” Journal of Statistical Software 64 (4): 1–34. \doi{10.18637/jss.v064.i04}
}
