% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Prevalence}
\usage{
prevalence(x)
}
\arguments{
\item{x}{number of parasites by host}
}
\value{
numeric value in \%.
}
\description{
Prevalence of parasites according to Bush et al. (1997).
}
\examples{
attach(birdticks)
prevalence(ticks)
}
\references{
Bush, Albert O., Kevin D. Lafferty, Jeffrey M. Lotz, and Allen W. Shostak. 1997. “Parasitology Meets Ecology on Its Own Terms: Margolis et al. Revisited.” The Journal of Parasitology 83 (4): 575. \doi{10.2307/3284227}
}
