% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_humidity_bulb.R
\name{relative_humidity_bulb}
\alias{relative_humidity_bulb}
\title{Relative humidity - Psycrometer}
\usage{
relative_humidity_bulb(dry_bulb, wet_bulb)
}
\arguments{
\item{dry_bulb}{dry bulb temperature in Celsius degree}

\item{wet_bulb}{wet bulb temperature in Celsius degree}
}
\description{
Relative humidity with wet and dry bulbs from psycrometer device according to Silveira-Neto et al. (1976).
}
\examples{
relative_humidity_bulb(27,22)
}
\references{
Silveira-Neto, Sinval, Octávio Nakano, Décio Barbin, and Nilson Augusto Villa Nova. 1976. Manual de Ecologia dos Insetos. Piracicaba, SP: Editora Agronômica Ceres LTDA.
}
