% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation_deficit.R
\name{saturation_deficit}
\alias{saturation_deficit}
\title{Saturation deficit}
\usage{
saturation_deficit(rh, t)
}
\arguments{
\item{rh}{relative humidity in \%}

\item{t}{temperature in Celsius degree}
}
\description{
water stress - saturation deficit using temperature and humidity parameters. Formula according to Randolph and Storey (1999).
}
\examples{
saturation_deficit(80, 27)
}
\references{
Randolph, Sarah E., and Katie Storey. 1999. “Impact of Microclimate on Immature Tick-Rodent Host Interactions (Acari: Ixodidae): Implications for Parasite Transmission.” Journal of Medical Entomology 36 (6): 741–48. \doi{10.1093/jmedent/36.6.741}
}
