% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation_deficit_bulb.R
\name{saturation_deficit_bulb}
\alias{saturation_deficit_bulb}
\title{Saturation deficit - Psycrometer}
\usage{
saturation_deficit_bulb(dry_bulb, wet_bulb)
}
\arguments{
\item{dry_bulb}{dry bulb temperature in Celsius degree}

\item{wet_bulb}{wet bulb temperature in Celsius degree}
}
\description{
saturation_deficit with wet and dry bulbs from psycrometer device according to Silveira-Neto et al. (1976).
}
\examples{
saturation_deficit_bulb(27,20)
}
\references{
Silveira-Neto, Sinval, Octávio Nakano, Décio Barbin, and Nilson Augusto Villa Nova. 1976. Manual de Ecologia dos Insetos. Piracicaba, SP: Editora Agronômica Ceres LTDA.
}
