% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{Rainy/dry season}
\usage{
season(date, start_rainy = "09-01", start_dry = "05-01")
}
\arguments{
\item{date}{date to classify}

\item{start_rainy}{Date of start rainy season ("month-day"). Default value consider 1st September.}

\item{start_dry}{Date of start dry season("month-day"). Default value consider 1st May.}
}
\value{
Character object classifying the target day as belonging to "dry season" or "rainy season".
}
\description{
Classification of seasons considering rainy and dry seasons.
}
\examples{
season("2021-05-23")
season("2020-03-06", start_rainy = "05-22", start_dry = "03-04")
}
\references{
Wolda, H. 1988. “Insect Seasonality: Why?” Annual Review of Ecology and Systematics 19 (1): 1–18. \doi{10.1146/annurev.es.19.110188.000245}.

“Köppen Climate Classification”. In Wikipedia, 1st November 2021. \url{https://en.wikipedia.org/w/index.php?title=K\%C3\%B6ppen_climate_classification&oldid=1053095128}.
}
