% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_mean.R
\name{var_mean}
\alias{var_mean}
\title{Variance/mean ratio}
\usage{
var_mean(x)
}
\arguments{
\item{x}{number of parasites by host}
}
\description{
Aggregation Index - Variance/mean
}
\examples{
attach(birdticks)
var_mean(ticks)
}
